+++ topic/src/main/java/tech/ydb/topic/impl/TopicClientImpl.java
@@ -13,10 +13,10 @@ import java.util.concurrent.Executors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import tech.ydb.core.Operations;
 import tech.ydb.core.Result;
 import tech.ydb.core.Status;
 import tech.ydb.core.grpc.GrpcRequestSettings;
+import tech.ydb.core.operation.OperationUtils;
 import tech.ydb.core.settings.BaseRequestSettings;
 import tech.ydb.core.utils.ProtobufUtils;
 import tech.ydb.topic.TopicClient;
@@ -85,7 +85,7 @@ public class TopicClientImpl implements TopicClient {
     @Override
     public CompletableFuture<Status> createTopic(String path, CreateTopicSettings settings) {
         YdbTopic.CreateTopicRequest.Builder requestBuilder = YdbTopic.CreateTopicRequest.newBuilder()
-                .setOperationParams(Operations.createParams(settings))
+                .setOperationParams(OperationUtils.createParams(settings))
                 .setPath(path)
                 .setRetentionStorageMb(settings.getRetentionStorageMb())
                 .setPartitionWriteSpeedBytesPerSecond(settings.getPartitionWriteSpeedBytesPerSecond())
@@ -121,7 +121,7 @@ public class TopicClientImpl implements TopicClient {
     @Override
     public CompletableFuture<Status> alterTopic(String path, AlterTopicSettings settings) {
         YdbTopic.AlterTopicRequest.Builder requestBuilder = YdbTopic.AlterTopicRequest.newBuilder()
-                .setOperationParams(Operations.createParams(settings))
+                .setOperationParams(OperationUtils.createParams(settings))
                 .setPath(path)/*
                 .putAllAttributes(settings.getAttributes())
                 .setMeteringMode(toProto(settings.getMeteringMode()))*/;
@@ -217,7 +217,7 @@ public class TopicClientImpl implements TopicClient {
     @Override
     public CompletableFuture<Status> dropTopic(String path, DropTopicSettings settings) {
         YdbTopic.DropTopicRequest request = YdbTopic.DropTopicRequest.newBuilder()
-                .setOperationParams(Operations.createParams(settings))
+                .setOperationParams(OperationUtils.createParams(settings))
                 .setPath(path)
                 .build();
         final GrpcRequestSettings grpcRequestSettings = makeGrpcRequestSettings(settings);
@@ -227,7 +227,7 @@ public class TopicClientImpl implements TopicClient {
     @Override
     public CompletableFuture<Result<TopicDescription>> describeTopic(String path, DescribeTopicSettings settings) {
         YdbTopic.DescribeTopicRequest request = YdbTopic.DescribeTopicRequest.newBuilder()
-                .setOperationParams(Operations.createParams(settings))
+                .setOperationParams(OperationUtils.createParams(settings))
                 .setPath(path)
                 .build();
         final GrpcRequestSettings grpcRequestSettings = makeGrpcRequestSettings(settings);